+++ RELEASE_NOTES.md
@@ -31,8 +31,9 @@ RELEASE NOTES
 Logtalk compiler and runtime
 ----------------------------
 
-* CHANGED: The built-in object `logtalk` is now provided in source code instead
-of being included in the compiler as a pre-compiled object.
+* CHANGED: The built-in object `logtalk` and the pseudo-object `user` are
+now provided in source code instead of being included in the compiler as
+pre-compiled objects.
 
 * CHANGED: The format of the execution context terms is now defined on the
 `logtalk` built-in object.